refactor(diagnostics): split OTEL content helpers
